Add new DefHook and AddHook functions
This commit adds two new functions to Package.pm called `DefHook` and
`AddHook`.

These allow to define hooks for existing macros. A hook can work like a
Primitive, a Macro or a Constructor. However, instead of only
performing the specified actions, a hook also runs the existing implementation.
The primary use-case is to introduce side-effects to an existing macro.

The `DefHook` function implements creating and registering a new control
sequence that perfoms a user-given side-effect, whereas the `AddHook`
function only registers an existing control-sequence as a hook for a
second already-existing control sequence.

The implementation of `AddHook(macro, hookmacro)` roughly corresponds to
the following TeX, but can also handle arguments passed to each of the
macros.

\let\originalmacro\macro
\def\macro{\hookmacro\originalmacro}