Set @filelist macro before reading file (fixes #857)
When loading a file through `InputDefinitions`, the @filelist macro was
only set after the file has finished loading. This results in the macro
being empty during the processing of a file.

This commit fixes the issue by setting the @filelist macro before
loading of the file. Furthermore, it adds a testcase to prevent this bug
from reappearing in the future.