Proper 'keyval' package support
This commit extends support for the keyval package by extending the
existing bindings to be able to read key-value definitions and the
associated defaults from raw macro definitions. Furthermore, it prepares
the implementation for support of extended packages, like 'xkeyval' and
friends.

Furthermore, it adds a new group of tests 'keyval'. For now, this
contains two tests, `keyvalinline` and `keyvalstyle`. The first of these
is testing the new functionality, the second one is an old test that was
moved out of the graphics package.