Add 'xkeyval' bindings
This commit adds bindings for the xkeyval package which work without
relying on the .sty file. These bindings do not support pre-setting keys
or using the built-in pointer system.

Furthermore, this commit refactors and extends the KeyVals class, to be
compatible with the new types of key-value pairs defined in the xkeyval
package. It also adds tests for all appropriate cases.