Fix missing runtime_class warning
When generating LaTeXML::MathGrammar, recent versions of
Parse::RecDescent were generating a warning about the missing
runtime_class option. This warning indiciated that the grammar may stop
working with future versions of the RecDescent package.

This commit adds the missing option to the call inside of Makefile.PL,
ensuring that the default runtime Parse::RecDescent will continue to be
used.