Bugfix for older Parse::RecDescent versions
Older versions of Parse::RecDescent did not support the third argument
to parser pre-compilation introduced in
b5aa3d04474538ea0d601d99e3e45d7a4df89aaa. This caused installation
problems for LaTeXML on older systems.

This commit fixes the problem by introducing a fallback that does not
use the third argument. This only gets triggered when the MathGrammar is
not created.