Use library for maketests instead of system()
Oddly, there were two `gen_xml` methods. The one being used was
shelling out to `latexml`. This doesn't work when working
with a development environment where `latexml` isn't on the path.

The other `gen_xml` method uses the library, which works with
local development environments. A nice side-effect is this is
the same approach used by `make test`, which seems more
resilient somehow.