Turn locators into an object
Previously, locators were strings. This made it difficult to work with
them in a programatic setting, or to extend them to contain source
ranges instead of points.

This commit updates the code to instead turn locators into a new
LaTeXML::Common::Locator class. It furthermore refactors the code, so
that the stringification of the locators only occurs upon output to the
user.