Add basic range support to Locator class
This commit adds basic support for locators representing ranges instead
of points to the LaTeXML::Common::Locator. It returns ranges for the
following two cases:
* LaTeXML::Common::Mouth records a locator representing the range of the
current token.
* LaTeXML::Common::WhatsIt records a locator that represents the range
of a body should it be present. Otherwise, it returns a point locator.

This commit also updates the behaviour of the stringify and toString
methods accordingly, and furthermore adds a new toAttribute method to
turn these locators into XPointer form for use within an XML Attribute.