Make Mouth Source '' for literal strings
Previously, we updated the source property of mouths to be undef in case of
anonymous strings.

This commit updates the previous behaviour and sets the mouth source to
be '' (the empty string) in case of a literal string.  This makes it
easier to see the difference between anonymous string inputs, which are
created on the fly by various bindings, and literal strings which can be
passed as valid inputs to LaTeXML.