Revamp the Travis Tests
This commit completly overhauls the way Travis tests are done.
Previously, these were used by installing appropriate versions of perl
and tex into the testing environment and then running a script.

This commit introduces a scheme for building appropriate docker test
runtime images. This allows for a lot more flexibility and reliable when
running tests. It also makes the tests run faster, as it is no longer
neccessary to download, install and configure several dependencies.