Refactor test suite behaviour
This commit refactors the behaviour of the test suite. This primarily consists
of:

Ensuring that normal tests are named '*.t' and extensive tests are named
'*.tt'. Furthermore, allow tests either directly in the 't/' directory or one
level deep.  We adapt 'Makefile.PL' so that 'make test' runs the normal tests
and 'make extratest' runs the extensive tests. Furthermore, 'make fulltest'
runs both tests; the Travis Tests are adapted to run this command.

Furthermore, when running tests, sort them alphabetically so that e.g. test
files in 't/000_unit/' get run before '005_tokenize.t'. All test files are now
prefixed with a three-digit-number (as opposed to a two-digit-number
previously).