verbatim fixes cascade
Surprising usage like "{\verb x }" led to realizing that
the skipping of whitespace after a control sequence doesn't
happen till reading the next token.  "\verb" immediately
changes the catcodes of space!
But isEOL queries whether the next token read would
be an EOL, so has to be skipping_spaces aware.
This led to rethinking how "\read" determines end of line;
noticing that "\futurelet" is NOT expandable

* Defer skipping spaces until next readToken
* Peak ahead in isEOL (!)
* fix how \read determines EOL
* make \futurelet a primitive, NOT expandable
* fixup \verb handling of space,
* in \verb defer EndSemiverbatim till after digestion, so visible space works
* fix \@makeother to deal with slashed chars
* fix \@verbatim to work standalone (but in group)
* remake test cases to match latex more closely
* enhance test cases to cover more usecases

The peak ahead in isEOL is bothersome!