Update CI to use Github Actions
Previously, the CI tests that were run after every commit used Travis
CI. Recent changes introduced a limit on the maximum time a build could
run for. LaTeXML easily exceeds this limit.

In order to keep the tests running, this commit moves the tests to
GitHub Actions. GitHub Actions is very similar to Travis except that it
is offered by GitHub directly and does not have a limit on Open Source
projects.

Because of previous work to move the testing environment into a Docker
Image, this move does not require many code changes. Furthermore, this
does not change which tests are run. In particular the versions of perl
and texlive being tested remain the same.