Remove the new timeout fatal tests.
It is a bit hard to stabilize the timeouts, and for a good reason - the stopping point is nondeterministic and depends on the specific machine's CPU and load.
So avoid being too precise with those tests - we already have one daemon test that ensures the timeout does indeed work, which should suffice.