BiBTeX/Runtime/Names: Handle "" better
Previously, the empty string was treated to contain exactly one name,
the empty name. This caused problems in specific .bst files that relied
on a name count eventually becoming 0.

This commit updates the behavior, and makes num.names$ and splitNames
return 0 and the empty array, respectively, for the empty input string.