ci: Use TeXLive 2018 - 2021, Perl 5.26 - 5.34
This commit updates the versions of TeXLive and Perl being tested to the
following combinations. These are derived from commonly used Linux
distributions and the version(s) of software they come with by default.

- TeXLive 2021 / Perl 5.34 (Archlinux, Fedora 35)
- TeXLive 2020 / Perl5.32 (Fedora 33-34, Debian 11, Ubuntu 21.04-21.10)
- TeXLive 2019 / Perl 5.30 (Ubuntu 20.04)
- TeXLive 2018 / Perl 5.28 (Debian 10)
- No TeXLive / Perl 5.26 - 5.34

This commit furhermore updates the GitHub workflow to only show up as
failed when a perl-only test fails.