Add 'role' attribute to 'ltx:resource'
This commit adds a new attribute to 'ltx:resources' that can be used to
indicate if the selected resource carries semantics or is simply
included by default.

This attribute is also added as a new option to RequireResource in
Package.pm, which only passes the attribute through to the xml.