Allow \bibliography to include '.bib'
Previously, when the argument to `\bibliography` ended in `.bib`,
MakeBibliography would fail to find the requested bibliography file.
This was cause by it adding an extra '.bib' extension, even though there
already was one.

This commit updates the behavior to allow MakeBibliography to find the
file, regardless if there is a '.bib' extension present or not. Special
care is taken for future compatibility with biber, which allows
bibliographies to have any extension.

Fixes #1605.