Rework APPVEYOR-specific code
This commit reworks code within LaTeXML that was specific to APPVEYOR CI
tests. Instead of having a single 'APPVEYOR' environment variable, it is
split into two new ones, and renamed more appropriatly:

'LATEXML_TEST_TEXLIVE=$YEAR', when present, pretends for the tests that
a specifc TeXLive version is installed without dynamically testing for
it. When `LATEXML_TEST_TEXLIVE=0` is provided, pretends that no texlive
is installed.

'LATEXML_KPSEWHICH_MIKTEX_ADMIN', when provided, appends a
`--miktex-admin` flag to all invocations of kpsewhich. This is useful
for windows installations where latexml is run as a superuser and MikTeX
requires this flag.

This commit also sets the above environment variables for the windows
tests.