Add Windows CI tests (#1612)
* Add Tests for Windows

This commit adds CI testing for windows using Strawberry Perl and no
TeXLive installation.

* Rename CI -> Linux

This commit renames the previous CI workflow to 'Linux' for consistency
with the new Windows CI.

* Expand Windows Tests to include MiKTeX

This commit extends the Windows CI Tests to include tests with MiKTeX.

* Rework APPVEYOR-specific code

This commit reworks code within LaTeXML that was specific to APPVEYOR CI
tests. Instead of having a single 'APPVEYOR' environment variable, it is
split into two new ones, and renamed more appropriatly:

'LATEXML_TEST_TEXLIVE=$YEAR', when present, pretends for the tests that
a specifc TeXLive version is installed without dynamically testing for
it. When `LATEXML_TEST_TEXLIVE=0` is provided, pretends that no texlive
is installed.

'LATEXML_KPSEWHICH_MIKTEX_ADMIN', when provided, appends a
`--miktex-admin` flag to all invocations of kpsewhich. This is useful
for windows installations where latexml is run as a superuser and MikTeX
requires this flag.

This commit also sets the above environment variables for the windows
tests.

* site: Workaround for failing MiKTeX installation

This commit documents the workaround for a failing installation when
running as an administrator under Windows and using MiKTeX.

* Install "complete" MiKTeX package set

This commit updates the Windows CI Test involving MiKTeX to install the
complete set of LaTeX packages. It also pretends to be TeXLive 2021 for
the tests.

* Use 'allow_failure' in Windows CI

This commit updates the 'allow_failure' behavior in the Windows CI to
match the behavior on the Linux CI. In particular, we allow Windows CI
with MiKTeX to fail, and the job to still pass.