beamer: Add new 'BeamerSquared' parameter type
Previously, the tests for [fragile] slides were incorrect and broken.
This was caused by a parameter of type [BeamerAngled] being followed by
another argument in square brackets.

This commit fixes the issue by adding a new 'BeamerSquared' parameter
type. This parameter type attempts to read an opening '[' and '<'. If it
finds both, it reads until appropriate closing brackets. If it does not,
it unreads the input. This parameter type effectively ignores non-angled
square bracket arguments.