Initial beamer.cls binding (#1652)
* schema: Add frame element and update slide element

This commit updates the schema to work well with beamer. In particular,
it adds a new 'frame' element and updates the existing 'slide' element.

* Initial beamer support

This commit adds initial support for beamer.cls.

* beamer: Inline beamer_support_{pre,post}load.ltxml

This commit inlines the two files beamer_support_{pre,post}load.ltxml
directly into beamer.cls.ltxml.

* beamer: Add RequirePackage('article.cls')

Previously, a large part of the binding in beamer was copied from
'article.cls.ltxml'. Instead of inlining a copy, instead add a
RequirePackage() call.

* beamer: Replace SquareAngled => [Angled]

This commit removes the SquareAngled ParameterType by the '[Angled]'
parameter type, removing the extra code.

* beamer: Fix broken article.cls load

* beamer: Update schema to allow slide or sequence

This commit updates the schema for (beamer) slides. First it renames
'frame' to the 'slidegroup'. Then it allows either a single slide, or a
group of slides in places where previously only a group of slides was
allowed.

This commit also updates the beamer binding and tests accordingly.

* schema: slideorsequence => slide | slidesequence

As requested by @brucemiller remove the `slideorschema` alias and replace
it by a literal `slide | slidesequence`.

* schema: Make slide overlay text

This commit makes the 'overlay' attribute for slides any text

* beamer: Add missing dependencies

* beamer: Ignore \setbeamersize

* schema: Remove unusued 'slidesequence.body.class'

The 'slidesequence.body.class' only contains a single element and is not
used anywhere in the schema, making it unneccessary. This commit removes
it from the schema.

* beamer: Add missing 'return'

This commit adds a missing 'return;' statement to avoid accidentally
returning digested objects.

* beamer: Run latexmllint

* beamer: Rename t/beamer => t/slides

This commmit renames the beamer tests from t/beamer to t/slides after a
suggestion from @brucemiller.

* beamer: Rename 'Angled' => 'BeamerAngled'

This commit renames the new 'Angled' parameter type to 'BeamerAngled' to
make clear it belongs to beamer and avoid conflicts with other angled
option types.

* beamer: Avoid loading begshi

* beamer: Add new 'BeamerSquared' parameter type

Previously, the tests for [fragile] slides were incorrect and broken.
This was caused by a parameter of type [BeamerAngled] being followed by
another argument in square brackets.

This commit fixes the issue by adding a new 'BeamerSquared' parameter
type. This parameter type attempts to read an opening '[' and '<'. If it
finds both, it reads until appropriate closing brackets. If it does not,
it unreads the input. This parameter type effectively ignores non-angled
square bracket arguments.

* beamer: Combine all tests into one

This commit combines all unit tests into a single test to save on the
overall number of tests.

* t/slides: Add a test for the classic slide package