Use optional argument to pagebreak
Examines the optional argument to `\pagebreak[]`, and outputs the
XML only when the argument is 3 or 4.  (The alternative would be
to end up with vertical spacing equal to `(argument/2)em`, but this
would require more complicated XML and also changing the CSS, not
to mention still needing to decide whether to break the page in
epub.)

Fixes #1802.