{La,}TeX.pool: Delay frontmatter digestion
As mentioned in brucemiller/LaTeXML#1929, native (La)TeX does not expand
arguments to '\title' until it is inserted into the document. LaTeXML on
the other hand digests the argument as soon as '\title' is invoked. This
causes differences in certain documents.

This commit updates the behaviour to only digest the argument at a later
point and store a Tokens() object in $STATE instead. Furthermore it
updates the behaviour of \maketitle to supported nested frontmatter.

Fixes brucemiller/LaTeXML#1929.