Split FrontMatter collection, digestion, insertion
This commit further tweaks the handling of frontmatter. Previously, it
was first collected, and then on an 'afterOpen' handle on 'ltx:document'
digested and promptly inserted into the document.

This is slightly different to the way native LaTeX handles, which first
collects frontmatter in specific macros. Then during expansion of
\maketitle it digests and inserts the stored frontmatter into the
document.

This commit makes LaTeXML more closely resemble the behaviour of LaTeX
by ensuring that digestion happens when '\maketitle' is called. The
insertion of frontmatter into the document is still handled seperatly
during the afterOpen call.

To also be able to handle documents without a '\maketitle' the digestion
phase is implemented so that it can safely be triggered repeatedly. It
is also triggered right before inserting frontmatter into the document.