latexmlc: Only unlink logfile when regular
As noted in #2069, `latexmlc` and friends were unlinking and re-creating
logfiles. This was not a problem in the regular case, however on certain
system when `make test` is run with root rights, this caused `/dev/null`
to be deleted and re-created as a regular file.

This commit updates the behavior of `latexmlc` to only unlink the
logfile if it is a regular file.

Fixes #2069.