Schema changes: More consistent naming and use of "block" elements, which are typically (vertical) blocks of block, para orsectional material (according to the level) for layout purposes.
* block can contain Block.model allowed in Block.class
* logical-block can contain Para.model allowed in Para.class
* sectional-block can contain sectional material, allowed in sections
Each has an inline- variant which is in Misc.class.

NOTE: that inline-para has been renamed inline-logical-block!
NOTE: that the content model of table cells (ltx:td) has been changed to Inline.model; if it needs to contain block level material, it needs to be contained in an appropriate inline-xxx element.