Support for dumping plain.tex, latex.ltx (#2525)
* Support for compact, preparsed dumps of latex.ltx, plain.tex formats for speed.
A format file is read in, and all changes to State are then Dumped to an external
file that can be quickly read in.

The normal binding file (LaTeX.pool.ltxml, plain.tex.ltxml) has been partitioned
into 3 components:
  * bootstrap : essential LaTeXML bindings necessary for processing the format
  * base or dump : the bulk of the format, (primarily macros & registers)
  * constructs : overlay's the definitions from main with LaTeXML-specific bindings,
    usually constructors, supporting document structure & semantics.

In its current form, the dump is a (somewhat) compact, text file encoding
the contents of the State tables, and can only contain macros or primitive values.
Constuctors and perl subroutines (which have no direct equivalent in TeX) are NOT stored.

An analog to iniTeX is provided by the --init option to latexml

* Lock \TeX,\LaTeX,\LaTeXe definitions so plain,latex.ltx don't replace ours

* Move \MakeUppercase and friends to constructs, so that we get consistent modern definition for testcases (ignore \uccode,\lccode; don't change math case)

* Make \nointerlineskip, \offinterlineskip macros so they expand away; update test cases

* Don't assume all \input files should be reloaded

* For dumping formats Use --destination with --init

* Provide more convenient API to load (possibly dumped) format files

* Add a target 'formats' to the Makefile which creates the dumps and puts into the blib directory (so, ready to run or install)